:+1: add user routing

+++ src/v1/controllers/user.py
@@ -6,7 +6,8 @@ from typing import List
 from database.base import User
 from database.session import with_session
 from models.user import UserORM
-from schemas import UserCreateSchema, UserSchema
+from aws_lambda_powertools.event_handler.exceptions import NotFoundError, BadRequestError
+from schemas import UserCreateSchema, UserSchema, UserUpdateSchema
 from sqlalchemy.orm.session import Session
 
 
@@ -21,6 +22,10 @@ class UserController:
 
     @with_session
     def create_one(self, session: Session, user_data: UserCreateSchema) -> tuple[UserSchema, int]:
+        if self.users.exists(db=session, user_id=user_data.id):
+            raise BadRequestError("userId already in use")
+        if self.users.find_by_email(db=session, email=user_data.email):
+            raise BadRequestError("email already in use")
         user = self.users.create_one(db=session, user_data=user_data)
         return UserSchema(**user.serializer()), HTTPStatus.CREATED.value
 
@@ -30,11 +35,22 @@ class UserController:
         return UserSchema(**user.serializer())
 
     @with_session
-    def update_one(self, session: Session, user_id: str, user_data: UserCreateSchema) -> UserSchema:
+    def find_one_or_404(self, session: Session, user_id: str) -> UserSchema:
+        if not self.users.exists(db=session, user_id=user_id):
+            raise NotFoundError
+        user = self.users.find_one(db=session, user_id=user_id)
+        return UserSchema(**user.serializer())
+
+    @with_session
+    def update_one(self, session: Session, user_id: str, user_data: UserUpdateSchema) -> UserSchema:
+        if not self.users.exists(db=session, user_id=user_id):
+            raise NotFoundError
         user = self.users.update_one(db=session, user_id=user_id, user_data=user_data)
         return UserSchema(**user.serializer())
 
     @with_session
     def delete_one(self, session: Session, user_id: str) -> None:
+        if not self.users.exists(db=session, user_id=user_id):
+            raise NotFoundError
         self.users.delete_one(db=session, user_id=user_id)
         return None
